don't exclude EI, UHC when looking for PDX=P
Though the CDM spec says "Relevant only on IP and IS encounters," table
Table IVE. Principal Diagnoses for Institutional Encounters of the EDC
report includes a row for EI as well.

We raised https://github.com/CDMFORUM/CDM-ERRATA/issues/19 about this.

Meanwhile, we have been counting only Epic billing diagnoses as
discharge diagnoses when looking for PDX. The
`\Diagnosis\Billing\Primary\` modifier used in UHC should be counted as
well. This modifier is also used in IDX, which will raise our count of
"For ED, AV, and OA encounter types, mark as X=Unable to Classify." but
that seems worthwhile.